requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
pydantic
deepagents
langchain
langchain-mcp-adapters